+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- `Node::drain` method to remove nodes by the predicate without memory allocations.
+
 ### Changed
 - Always add a space after ArcTo flags during the path writing.
 - SVG and non-SVG attributes now stored in the same container and not separately.